Add files via upload
As a part of my learning, I gathered data of a reputed retailer company. I used SQL to get the insights from the data. 
In the above mentioned file, I myself have made questions and written the code accordingly and pasted the output.
In this project, I had mainly concentrated on revenue and logistics. I had also discussed the suggestions based on the output.
Did identified the trend on the orders and revenue. Please do check out my work. Thanks.